move remaining cli tests to tests/e2e
