+++ tests/e2e/auth/vesting/client/testutil/cli_test.go
@@ -10,12 +10,13 @@ import (
 	"github.com/stretchr/testify/suite"
 
 	"github.com/cosmos/cosmos-sdk/testutil/network"
-	"github.com/cosmos/cosmos-sdk/x/mint/testutil"
+	testutil2 "github.com/cosmos/cosmos-sdk/x/auth/vesting/client/testutil"
+	"github.com/cosmos/cosmos-sdk/x/auth/vesting/testutil"
 )
 
 func TestIntegrationTestSuite(t *testing.T) {
 	cfg, err := network.DefaultConfigWithAppConfig(testutil.AppConfig)
 	require.NoError(t, err)
 	cfg.NumValidators = 1
-	suite.Run(t, NewIntegrationTestSuite(cfg))
+	suite.Run(t, testutil2.NewIntegrationTestSuite(cfg))
 }
