+++ tests/e2e/feegrant/client/testutil/cli_test.go
@@ -7,6 +7,7 @@ import (
 	"testing"
 
 	"github.com/cosmos/cosmos-sdk/testutil/network"
+	clienttestutil "github.com/cosmos/cosmos-sdk/x/feegrant/client/testutil"
 	"github.com/cosmos/cosmos-sdk/x/feegrant/testutil"
 
 	"github.com/stretchr/testify/require"
@@ -17,5 +18,5 @@ func TestIntegrationTestSuite(t *testing.T) {
 	cfg, err := network.DefaultConfigWithAppConfig(testutil.AppConfig)
 	require.NoError(t, err)
 	cfg.NumValidators = 3
-	suite.Run(t, NewIntegrationTestSuite(cfg))
+	suite.Run(t, clienttestutil.NewIntegrationTestSuite(cfg))
 }
