+++ tests/e2e/group/client/testutil/cli_test.go
@@ -7,7 +7,8 @@ import (
 	"testing"
 
 	"github.com/cosmos/cosmos-sdk/testutil/network"
-	"github.com/cosmos/cosmos-sdk/x/slashing/testutil"
+	clienttestutil "github.com/cosmos/cosmos-sdk/x/group/client/testutil"
+	"github.com/cosmos/cosmos-sdk/x/group/testutil"
 
 	"github.com/stretchr/testify/require"
 	"github.com/stretchr/testify/suite"
@@ -16,6 +17,6 @@ import (
 func TestIntegrationTestSuite(t *testing.T) {
 	cfg, err := network.DefaultConfigWithAppConfig(testutil.AppConfig)
 	require.NoError(t, err)
-	cfg.NumValidators = 1
-	suite.Run(t, NewIntegrationTestSuite(cfg))
+	cfg.NumValidators = 2
+	suite.Run(t, clienttestutil.NewIntegrationTestSuite(cfg))
 }
