+++ tests/e2e/mint/client/testutil/cli_test.go
@@ -6,16 +6,17 @@ package testutil
 import (
 	"testing"
 
-	"github.com/cosmos/cosmos-sdk/testutil/network"
-	"github.com/cosmos/cosmos-sdk/x/group/testutil"
-
 	"github.com/stretchr/testify/require"
 	"github.com/stretchr/testify/suite"
+
+	"github.com/cosmos/cosmos-sdk/testutil/network"
+	clienttestutil "github.com/cosmos/cosmos-sdk/x/mint/client/testutil"
+	"github.com/cosmos/cosmos-sdk/x/mint/testutil"
 )
 
 func TestIntegrationTestSuite(t *testing.T) {
 	cfg, err := network.DefaultConfigWithAppConfig(testutil.AppConfig)
 	require.NoError(t, err)
-	cfg.NumValidators = 2
-	suite.Run(t, NewIntegrationTestSuite(cfg))
+	cfg.NumValidators = 1
+	suite.Run(t, clienttestutil.NewIntegrationTestSuite(cfg))
 }
